Add tarball for all source files
I am uploading the tarball that includes all source files (that is, .java files) to this repo as this makes it easier to download this project's source code from this repo. Note that the tarball in this specific repo does not include any of the images, .class files, or .html files need for actually running the game.